ch03_openai/requirements.txt
openai==1.30.5
python-dotenv==1.0.1
